Cha Cha Cha, Changes
I have narrowed down the selection of buttons to nine to keep the moral
(read - quality) up, as well as rewrite the buttons to be anchor links f
or easier styling on your end, although you could easily change it to su
it your needs. Cleaned some dusty code and chucked a couple of things, n
ot much life changing.